Log addArtifacts errors
